LayoutTests/editing/inserting/insert-paragraph-separator-tab-span-expected.txt

This test ensures CyberKit avoids cloning Apple tab span when inserting a paragraph separator.
Only tab should be inside a Apple tab span in the following tests.

last visible position:
| <span>
|   class="Apple-tab-span"
|   style="white-space:pre"
|   "	"
| <div>
|   "hello world CyberKit <#selection-caret>"

first visible position:
| <div>
|   <br>
| "hello world CyberKit <#selection-caret>"
| <span>
|   class="Apple-tab-span"
|   style="white-space:pre"
|   "	"

before tab span:
| "hi, "
| <div>
|   "hello world CyberKit <#selection-caret>"
|   <span>
|     class="Apple-tab-span"
|     style="white-space:pre"
|     "	"
|   " rocks"

after tab span:
| "hi, "
| <span>
|   class="Apple-tab-span"
|   style="white-space:pre"
|   "	"
| <div>
|   "hello world CyberKit <#selection-caret>rocks"
